README and LICENSE
This is an initial commit with some boilerplate code and stuff.

Yes, you can think of this as a pet project because I got bored of
not writing Rust code for a while. Or maybe I'm bored with mundane
tasks of maintaining other language wrappers of Themis...

I'm the original author of the Rust-Themis wrapper and it has taught
me a lot about interfacing C code from Rust and interacting with C
build systems. This time I'm going to learn the reverse: how to
interact with Rust code from C, provide C bindings, etc.