Use util/read_symbols.go instead of goblin (#3)
It turns out that BoringSSL already includes a utility for reading
the symbols of produced libraries. Since we depend on Golang for
the build system to work, we can just use this tool instead of
pulling in some Rust dependency.

On my machine using this tools leads to 2x speedup for clean builds:
down to around 30 seconds from around 1 minute. Plus we get rid of
a dependency and have fewer code to maintain. I'd take it any day.

Furthermore, the Rust code seems to be based on that utility.
I'd rather rely on BoringSSL core to be maintained.