Generate bindings with bindgen
Let's use bindgen to generate Rust bindings to the library we build.
We use the same approach and reuse the code from Google's Mundane [1]
project. I made some simplifications in the script, tweaked it for
Soter, and moved the whitelist into a separate file so that it's
easier to maintain.

The scripts expects to have both functions and non-functions (types
and global variables) to be listed in the symbol whitelist. However,
currently we don't have much to use. I have whitelisted a function
that we're certain to use and the BIGNUM structure which is very
likely to be used. We may also need to pass --whitelist-var later,
but bindgen complains about it if there are no actual variables
present so it has been removed.

[1]: https://github.com/google/mundane/tree/master/boringssl