Sanity checks
We're not really going to test the raw FFI bindings extensively,
but we should have a simple check to verify that everything seems
to be in place. We try calling some random function (he-he-he)
and see if the linking works correctly and the call is successful.
This is good enough. soter-boringssl crate will have more "real"
sanity tests for BoringSSL API.