Tweak static library names on Windows
UNIX systems generally name their static libraries "libfoo.a"
while native Windows toolchains produce "foo.lib" files.
Windows does not have issues with "libfoo.a", but that's what
CMake chooses by default for libcrypto and we have to use that,

Tweak the code to work with correct library names, and simplify
it a little bit.