Test across Linux, macOS, and Windows (#5)
Add a test matrix to the workflow to run in on all popular desktop
operating systems. Now I'm convinced that this code is portable.

It took me a while to figure out what exactly is missing right now
for each of GitHub Actions' platform, but now it kinda works and
I know exactly what needs to be done (despite lack of experience
with native Windows builds).

Linux machines have everything available out-of-the-box, they are
the best. They do not require any tweaks.

On macOS we do not have Rust toolchain installed, Thankfully, some
people have already made an action for that so we use it. Other
platforms have some recent stable distribution installed already,
so we can save 1-2 minutes of build time there.

Windows does not have Ninja and NASM installed, which are both
required by BoringSSL. For Ninja we have a third-party action and
NASM is available in Chocolatey repositories. It takes a while
to install, but such is life. Next, Windows machines have MSVC++
installed, but we have to explicitly configure the toolchain and
reexport some important environment variables so that CMake sees
and uses the correct compiler, runtime libraries, etc.

This covers three most important targets:

  - x86_64-apple-darwin
  - x86_64-pc-windows-msvc
  - x86_64-unknown-linux-gnu

Note that we do not test with x86_64-pc-windows-gnu which is kinda
important as well. It should be trivial to setup, but I can't be
bothered right now. Themis also supports various i686-* targets
on Windows and Linux, and a multitude of ARM architectures for
iOS and Android devices. It would be nice to add them later.

* Tweak static library names on Windows

UNIX systems generally name their static libraries "libfoo.a"
while native Windows toolchains produce "foo.lib" files.
Windows does not have issues with "libfoo.a", but that's what
CMake chooses by default for libcrypto and we have to use that.

Tweak the code to work with correct library names, and simplify
it a little bit.