Benchmark random byte generation
Because why not? Now I know that I can generate 2.7 GB/s on my machine.
Let's merge this commit that was lying around from 2020-10-06.

Add Criterion-based benchmarks that can be used with stable Rust
toolchain. This is more of an example and template for how to add
new benchmarks later.

Initially I wanted to add benchmark runs to CI too, but it is going
to take so much time to run them -- while not giving particularly
useful results due to all the noise that exists in CI pipelines
running in shared VMs. I'm fine with benchmarks being compilable
(verified by "cargo check --all") and runninng them on developer
machines as necessary.