Update bindgen flag set
Recent updates in Bindgen has changed the way it generates boilerplate
comment. Disable built-in generation with "--disable-header-comment"
but keep the old stable comment in the boilerplate we add.

Also, disable struct layout tests which are platform-dependent. This
would have been okay if Bindgen was used from build.rs on the developer
machines, but we check in the generated code to avoid having the users
to install Bindgen (or rather, compile it every time they compile this
project). Therefore we should not include non-portable code here.

Finally, automatically convert C's "size_t" into Rust "usize". Many APIs
of BoringSSL use "size_t" and it's bothersome to deal with this special
type all the time when there is a native Rust equivalent.