Update Bindgen script (#14)
* Update bindgen flag set

Recent updates in Bindgen has changed the way it generates boilerplate
comment. Disable built-in generation with "--disable-header-comment"
but keep the old stable comment in the boilerplate we add.

Also, disable struct layout tests which are platform-dependent. This
would have been okay if Bindgen was used from build.rs on the developer
machines, but we check in the generated code to avoid having the users
to install Bindgen (or rather, compile it every time they compile this
project). Therefore we should not include non-portable code here.

Finally, automatically convert C's "size_t" into Rust "usize". Many APIs
of BoringSSL use "size_t" and it's bothersome to deal with this special
type all the time when there is a native Rust equivalent.

* Improved whilelist and opaque types

Actually, since we can separate functions and types in the whitelist, we
can mention them separately on the command line too. This saves us from
some warnings.

Also, make all types we import opaque. We should not depend on the
layout or internal details of BoringSSL types, always using accessors
instead.

* Use /usr/bin/env trick for flexibility

Instead of hardcoding /bin/bash location, use the one from user's PATH.
This is generally good for portability, and gives the user some freedom
to run the script with a different shell.

* Autodetect crate version

Instead of forcing maintainers to type in the version to include into
generated bindings, just pick it up from the Cargo.toml file. However,
leave an option to specify the version explicitly if there is a need.
(Environment variables are fine too but they are more bothersome.)

* Update binding source code

Since we have converted all types to be opaque, BIGNUM now does not
expose any details, instead it has "u8" alignment for pointer types
to have accurate alignment as well.