Error type for BoringSSL
Instead of returning raw BoringSSL error codes (and duplicating their
treatment all around Soter), return a proper Result with an Error
enumeration, similar to the one used by Soter. In fact, they are almost
the same thing.

Now the error treatment can be performed in "soter-boringssl" crate
which knows exactly how each BoringSSL function behaves and what is its
convention for errors.

The error is now also parameterized by the Ok type so that it's possible
to return values with this Result. This will be important for various
constuctors and simple accessors.

There is also a ResultExt trait providing utility methods to convert
from integer error codes to Result instances. This trait replaces the
utility functions since methods make it easier to chain calls with
Result methods if necessary.