Mid-level safe EVP API in soter-boringssl
These are basically safe wrappers over the raw FFI calls. Most of them
are straightforward. Naming and calling conventions are kept as close
to the original as possible to simplify code review.

However, effort is applied to provide a safe interface: use RAII for
managing lifetimes, return errors with Result, accept slices instead of
separate buffers and their sizes.

EVP_DigestFinal_ex() is wacky. Initially I assumed that it checks the
out-buffer size via the given parameter and avoids overflow. But in fact
it expects the buffer to have sufficient size and the "size" parameter
is strictly output. I planned to have the size check in "soter" crate,
but it seems to be needed in "soter-boringssl" for it to provide a safe
interface.