High-level Hash computation API in soter
Hash function computation is implemented in the Hash struct. It's
mostly straightforward, but there are some design points that need
clarifications:

- Algorithm selection is dynamic, because

  a) BoringSSL EVP API is fundamentally dynamic;
  b) Soter C API is also dynamic.

Therefore there is no reason for static algorithm selection other than
convenience, which can be provided later. I'm not going to make it
prematurely optimized by introducing static dispatch with generics.

- Zero-allocation "finalise"

This is important for efficient code. However, it has a wanky interface
in return. For Rust users there is an easier "get" method which returns
a newly allocated vector. Soter C API is still going to use "finalise".

Also note that while "get" consumes Hash and prevents erroneous calls to
"write" and "finalise", "finalise" must leave the "Hash" object intact
because it will be deallocated later in Soter C API.

- Panics vs. Result errors

In some cases the API returns errors, but it most cases it will panic.
The reason is that the errors are not normally recoverable. The errors
here are not caused by user input or user actions, but rather by either

  a) Soter's incorrect usage of BoringSSL, triggering an error
     which should not happen;

  b) bugs in Soter's build system, disabling some algorithms
     which should be always available;

  c) bugs in the application programming: unfortuntely, we cannot make
     "finalise" 100% safe *and* efficient due to BoringSSL quirks;

  d) due to some interal issue in BoringSSL which neither Soter,
     nor the developers, nor the users have any control over.

That is, there are no business logic errors that are expected to be
handled in hash function computation. If you run into such panic in
production code, you're free to work around it by catching the panic.

The errors that *can* reasonably be handled by the application *are*
returned as Result errors.