SHA-2 benchmarks
Add some benchmarks to see if there is any noticeable performance hit
because of all additional checks in Rust code, how SHA-256 and SHA-512
compare to each other on typical hardware, and if there are anomalies.

There is some consistent performance drop, but it's not that significant
and I'm not really in the mood for optimizing the hell out of it.