Hash computation API (#16)
* Low-level EVP API in soter-boringssl-sys

Add functions necessary for hash computation to the whilelist and
bindings. Note that some of the EVP API is macro-based and as such
it is not accessible via Bindgen (currently, at least). We need
to use actual functions that have actual symbols.

* Mid-level safe EVP API in soter-boringssl

These are basically safe wrappers over the raw FFI calls. Most of them
are straightforward. Naming and calling conventions are kept as close
to the original as possible to simplify code review.

However, effort is applied to provide a safe interface: use RAII for
managing lifetimes, return errors with Result, accept slices instead of
separate buffers and their sizes.

EVP_DigestFinal_ex() is wacky. Initially I assumed that it checks the
out-buffer size via the given parameter and avoids overflow. But in fact
it expects the buffer to have sufficient size and the "size" parameter
is strictly output. I planned to have the size check in "soter" crate,
but it seems to be needed in "soter-boringssl" for it to provide a safe
interface.

* High-level Hash computation API in soter

Hash function computation is implemented in the Hash struct. It's
mostly straightforward, but there are some design points that need
clarifications:

- Algorithm selection is dynamic, because

  a) BoringSSL EVP API is fundamentally dynamic;
  b) Soter C API is also dynamic.

Therefore there is no reason for static algorithm selection other than
convenience, which can be provided later. I'm not going to make it
prematurely optimized by introducing static dispatch with generics.

- Zero-allocation "finalise"

This is important for efficient code. However, it has a wanky interface
in return. For Rust users there is an easier "get" method which returns
a newly allocated vector. Soter C API is still going to use "finalise".

Also note that while "get" consumes Hash and prevents erroneous calls to
"write" and "finalise", "finalise" must leave the "Hash" object intact
because it will be deallocated later in Soter C API.

- Panics vs. Result errors

In some cases the API returns errors, but it most cases it will panic.
The reason is that the errors are not normally recoverable. The errors
here are not caused by user input or user actions, but rather by either

  a) Soter's incorrect usage of BoringSSL, triggering an error
     which should not happen;

  b) bugs in Soter's build system, disabling some algorithms
     which should be always available;

  c) bugs in the application programming: unfortuntely, we cannot make
     "finalise" 100% safe *and* efficient due to BoringSSL quirks;

  d) due to some interal issue in BoringSSL which neither Soter,
     nor the developers, nor the users have any control over.

That is, there are no business logic errors that are expected to be
handled in hash function computation. If you run into such panic in
production code, you're free to work around it by catching the panic.

The errors that *can* reasonably be handled by the application *are*
returned as Result errors.

* SHA-2 test suite

Use some test vectors provided by NIST to verify that the
implementation seems to indeed compute SHA-256 and SHA-512.
There are some long tests thrown for completeness, you can
enable those with "--features long_tests". They take a couple
of seconds to run on modern hardware.

Aside from test vectors, verify other important points in the
implementation which are significant for the API.

* SHA-2 benchmarks

Add some benchmarks to see if there is any noticeable performance hit
because of all additional checks in Rust code, how SHA-256 and SHA-512
compare to each other on typical hardware, and if there are anomalies.

There is some consistent performance drop, but it's not that significant
and I'm not really in the mood for optimizing the hell out of it.